--- src/preferences/user.js
@@ -1,24 +1,0 @@
-//
-// export class UserPreferences {
-//     constructor() {
-//         this.preferences = JSON.parse(localStorage.getItem('userPreferences')) || {};
-//     }
-//
-//     set(key, value) {
-//         this.preferences[key] = value;
-//         this.save();
-//     }
-//
-//     get(key) {
-//         return this.preferences[key];
-//     }
-//
-//     save() {
-//         localStorage.setItem('userPreferences', JSON.stringify(this.preferences));
-//     }
-//
-//     clear() {
-//         localStorage.removeItem('userPreferences');
-//         this.preferences = {};
-//     }
-// }
