+++ src/storage/storage.js
@@ -1,0 +1,11 @@
+export const saveLocalStorage = (key, value) => {
+  localStorage.setItem(key, JSON.stringify(value));
+};
+
+export const getLocalStorage = (key) => {
+  return JSON.parse(localStorage.getItem(key));
+};
+
+export const clearLocalStorage = () => {
+  localStorage.clear();
+};
